netflow_producer/pyproject.toml
[project]
name = "netflow-producer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "kafka-python>=2.1.5",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic-settings>=2.9.1",
    "pydantic[email,timezone]>=2.11.3",
]
